ref: split routes file to api, services, repositories ans cache
